[BE/FEAT] MEMBER 몸무게 입력, 이력 조회 구현

See merge request s09-bigdata-recom-sub2/S09P22A403!96

+++ backend/fns/src/main/java/ssafy/fns/domain/member/service/dto/MemberResponseDto.java
@@ -14,8 +14,8 @@ public class MemberResponseDto {
 
     private String nickname;
     private Long age;
-    private Long height;
-    private Long weight;
+    private Double height;
+    private Double weight;
     private String gender;
     private Boolean isPublished;
